Move vec-slice-drop test
